Fixing part display name edition
