+++ src/OrchardCore.Modules/OrchardCore.ContentTypes/Editors/ContentPartSettingsDisplayDriver.cs
@@ -18,6 +18,7 @@ namespace OrchardCore.ContentTypes.Editors
                 model.Attachable = settings.Attachable;
                 model.Reusable = settings.Reusable;
                 model.Description = settings.Description;
+                model.DisplayName = settings.DisplayName;
 
                 return Task.CompletedTask;
             }).Location("Content");
@@ -32,6 +33,7 @@ namespace OrchardCore.ContentTypes.Editors
                 context.Builder.Attachable(model.Attachable);
                 context.Builder.Reusable(model.Reusable);
                 context.Builder.WithDescription(model.Description);
+                context.Builder.WithDisplayName(model.DisplayName);
             }
 
             return Edit(contentPartDefinition, context.Updater);
