+++ src/OrchardCore.Modules/OrchardCore.ContentTypes/ViewModels/ContentPartSettingsViewModel.cs
@@ -1,9 +1,10 @@
-﻿namespace OrchardCore.ContentTypes.ViewModels
+namespace OrchardCore.ContentTypes.ViewModels
 {
     public class ContentPartSettingsViewModel
     {
         public bool Attachable { get; set; }
         public bool Reusable { get; set; }
         public string Description { get; set; }
+        public string DisplayName { get; set; }
     }
 }
