+++ src/OrchardCore.Modules/OrchardCore.ContentTypes/Views/ContentPartSettings.Edit.cshtml
@@ -1,4 +1,13 @@
-﻿@model ContentPartSettingsViewModel
+@model ContentPartSettingsViewModel
+
+<fieldset class="form-group">
+    <div class="row col-md ">
+        <label asp-for="DisplayName">@T["Display Name"]</label>
+        <input asp-for="DisplayName" class="form-control" />
+        <span class="hint">@T["Name of the part as it will be displayed in screens."]</span>
+    </div>
+</fieldset>
+
 <fieldset class="form-group">
     <div class="row col-lg ">
         <label asp-for="Description">@T["Description"]</label>
