+++ src/OrchardCore.Modules/OrchardCore.ContentTypes/Views/Shared/DisplayTemplates/EditPartViewModel.cshtml
@@ -1,9 +1,10 @@
-﻿@model EditPartViewModel
+@using OrchardCore.ContentManagement.Metadata.Settings
+@model EditPartViewModel
 <div class="properties">
     <div class="related">
         <a class="btn btn-primary btn-sm" asp-route-action="EditPart" asp-route-id="@Model.Name">@T["Edit"]</a>
     </div>
-    <div>@Model.DisplayName</div>
+    <div>@Model.PartDefinition.DisplayName()</div>
     @if (!string.IsNullOrWhiteSpace(Model.Description))
     {
         <span class="hint">@Model.Description</span>
