+++ src/OrchardCore/OrchardCore.ContentManagement.Abstractions/Metadata/Models/ContentPartDefinition.cs
@@ -1,4 +1,4 @@
-﻿using Newtonsoft.Json.Linq;
+using Newtonsoft.Json.Linq;
 using System.Collections.Generic;
 using System.Linq;
