+++ src/OrchardCore/OrchardCore.ContentManagement.Abstractions/Metadata/Settings/ContentPartSettingsExtensions.cs
@@ -1,5 +1,7 @@
-﻿using OrchardCore.ContentManagement.Metadata.Builders;
+using System;
+using OrchardCore.ContentManagement.Metadata.Builders;
 using OrchardCore.ContentManagement.Metadata.Models;
+using OrchardCore.Mvc.Utilities;
 
 namespace OrchardCore.ContentManagement.Metadata.Settings
 {
@@ -30,9 +32,26 @@ namespace OrchardCore.ContentManagement.Metadata.Settings
             return builder.WithSetting(nameof(ContentPartSettings.Description), description);
         }
 
+        public static ContentPartDefinitionBuilder WithDisplayName(this ContentPartDefinitionBuilder builder, string description)
+        {
+            return builder.WithSetting(nameof(ContentPartSettings.DisplayName), description);
+        }
+
         public static string Description(this ContentPartDefinition part)
         {
             return part.Settings.ToObject<ContentPartSettings>().Description;
+        }
+
+        public static string DisplayName(this ContentPartDefinition part)
+        {
+            var displayName = part.Settings.ToObject<ContentPartSettings>().DisplayName;
+
+            if (String.IsNullOrEmpty(displayName))
+            {
+                displayName = part.Name.TrimEnd("Part");
+            }
+
+            return displayName;
         }
     }
 }
